requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
base36==0.1.1
certifi==2023.11.17
cffi==1.16.0
chacha20poly1305-reuseable==0.12.0
charset-normalizer==3.3.2
cryptography==41.0.7
frozenlist==1.4.1
h11==0.14.0
HAP-python==4.9.1
idna==3.6
ifaddr==0.2.0
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.9.10
psycopg2-binary==2.9.9
pycparser==2.21
PyQRCode==1.2.1
python-dotenv==1.0.0
requests==2.31.0
RPi.GPIO==0.7.1
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
zeroconf==0.131.0